moved template to plant card, removed some conflicting css
