+++ client/src/components/PlantCard/PlantCard.css
@@ -1,94 +1,85 @@
-/* .Card {
-    margin: 8px;
-} */
 
-/* .plant-imgurl {
-    margin-top: 20px;
-    margin-bottom: 20px;
-} */
+.plant-card-section {
+  /* background-color: #000e21; */
+  padding: 8px;
+  margin-top: 50px;
+  margin-bottom: 20px;
 
-.row {
-    /* width: 100%; */
-    /* display: flex; */
-    
+  min-height: 500px;
+  /* border: 3px solid grey; */
+  /* box-shadow: 6px 8px rgba(0, 0, 0, 0.4); */
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-template-rows: fit-content(1em);
+  grid-template-areas:
+    "header header header"
+    "mostrecent mostrecent mostrecent";
+  align-content: center;
+  overflow: hidden;
+}
+
+.recent-card-header {
+  grid-area: header;
+}
+/* this is being used on the plants page, leave this temporarily until you add the card there */
+.plant-card-plant {
+  grid-area: mostrecent;
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(325px, 1fr));
+}
+
+.most-recent {
+  grid-area: mostrecent;
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(325px, 1fr));
+  position: relative;
+  padding-top: 0px;
+  margin-top: 0px;
+}
+
+.plant-card-wrapper {
+  position: relative;
+}
+
+.plant-card-img {
+  border-radius: 2px;
+  box-shadow: 0 8px 16px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
+  display: grid;
+  height: 400px;
+  overflow: hidden;
+  opacity: 1;
+  z-index: 1;
+  position: relative;
+
+}
+
+.plant-card-img-overlay {
+  background-color: rgba(1, 1, 1, 0.5);
+  top: 0;
+  right:0;
+  opacity: 0;
+  position: absolute;
+  z-index: 2;
+  height: 400px;
+  overflow: hidden;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  font-size: 21px;
+  font-weight: 300;
+  cursor: pointer;
 }
 
 
-/* *, ::after, ::before {
-    box-sizing: border-box;
-} */
-.plant-details-sec {
-    
-    margin-top: 20px;
-    margin-bottom: 20px;
-    flex-wrap: nowrap;
-    /* width: 100%; */
-    /* flex-direction: row;
-    justify-content: space-between; */
-    /* align-items: baseline; */
-    /* justify-content: space-evenly; */
-    justify-content: space-evenly;
-    padding: 10px;
-    margin: 10px;
+.plant-card-img-overlay:hover {
+  opacity: 1;
+  z-index: 5;
+  color: whitesmoke;
+  justify-content: center;
 }
 
-.plant-img-left {
-    /* padding: 10px; */
-    margin-right: 20px;
-    margin-left: 20px;
-    
-}
 
-.plant-card-x {
-    /* max-width: 600px; */
-    min-width: 500px;
-    background-color: white;
-    width: 100%;
-    
-    
-}
-
-.plant-card-header {
-    background-color: #5F9EA0;
-    padding: 20px;
-    font-size: 20px;
-    font-weight: 600;
-    text-align: left;
-    /* width: 100%; */
-}
-
-.plant-card-details-group, .button-section-header {
-    text-align: left;
-}
-
-.plant-card-details {
-    list-style-type: none;
-    border-bottom: 0.03em solid lightgrey;
-    padding: 10px;
-    /* word-wrap: break-word; */
-    /* border-style: inset; */
-    /* opacity: 0.5; */
-}
-
-button.always {
-    /* display: none; Hidden by default */
-  position: fixed; /* Fixed/sticky position */
-  bottom: 20px; /* Place the button at the bottom of the page */
-  right: 100px; /* Place the button 30px from the right */
-  z-index: 99; /* Make sure it does not overlap */
-  background-color: grey; /* Set a background color */
-  color: white; /* Text color */
-  cursor: pointer; /* Add a mouse pointer on hover */
-  padding: 15px; /* Some padding */
-  border-radius: 10px; /* Rounded corners */
-  font-size: 18px; /* Increase font size */   
-}
-
-button.always:hover {
-    background-color: lightgrey;
-}
-
-.water-button {
-    border-radius: 4px; 
-    padding: 5px;
+.fa-ex1 {
+  justify-items: flex-end;
 }
