+++ client/src/components/RecentAdditions/RecentAdditions.css
@@ -14,9 +14,7 @@
   grid-template-rows: fit-content(1em);
   grid-template-areas:
     "header header header"
-    /* "mostrecent mostrecent mostrecent" */
-    "mostrecentindoor mostrecentindoor mostrecentindoor"
-    "mostrecentoutdoor mostrecentoutdoor mostrecentoutdoor";
+    "mostrecent mostrecent mostrecent";
   align-content: center;
   overflow: hidden;
 }
@@ -27,15 +25,8 @@
   text-align: left;
 }
 
-.most-recent-plants {
+.most-recent {
   grid-area: mostrecent;
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(325px, 1fr));
-  position: relative;
-}
-
-.most-recent-indoor {
-  grid-area: mostrecentindoor;
   display: grid;
   grid-template-columns: repeat(auto-fit, minmax(325px, 1fr));
   position: relative;
@@ -43,86 +34,14 @@
   margin-top: 0px;
 }
 
-.most-recent-outdoor {
-  grid-area: mostrecentoutdoor;
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(325px, 1fr));
+.more-plants-card {
   position: relative;
 }
 
-.plant-card-wrapper {
-  position: relative;
-}
-
-.plant-card {
-  border-radius: 2px;
-  box-shadow: 0 8px 16px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
-  display: grid;
-  height: 400px;
-  overflow: hidden;
-  opacity: 1;
-  z-index: 1;
-  position: relative;
-
-}
-
-/* .plant-card:hover {
-  -webkit-transform: scale(1, 1);
-  transform: scale(1.03, 1.03);
-  opacity: 1;
-  z-index: 5;
-  /* position: relative;
-} */
-
-.plant-card-2 {
+.more-plant-card {
   background-color: rgba(1, 1, 1, 0.5);
   top: 0;
   right:0;
-  opacity: 0;
-  position: absolute;
-  z-index: 2;
-  height: 400px;
-  overflow: hidden;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-size: 21px;
-  font-weight: 300;
-  cursor: pointer;
-}
-
-
-.plant-card-2:hover {
-  opacity: 1;
-  z-index: 5;
-  color: whitesmoke;
-  justify-content: center;
-}
-
-.plant-caption {
-  background-color: whitesmoke;
-  /* width:25vw; */
-}
-
-.name-img {
-  position: absolute;
-  background-color: cadetblue;
-  padding: 5px;
-  z-index: 10;
-}
-
-.see-more {
-  position: relative;
-}
-
-.plant-card-see-more {
-  background-color: rgba(1, 1, 1, 0.5);
-  top: 0;
-  right:0;
-  opacity: 1;
-  position: relative;
-  z-index: 1;
   height: 400px;
   overflow: hidden;
   width: 100%;
@@ -135,26 +54,19 @@
   color: white;
 }
 
-.plant-card-see-more-2 {
-  background-color: rgba(1, 1, 1, 0.5);
-  top: 0;
-  right:0;
+.more-plant-card-text {
+  opacity: 1;
+  position: relative;
+  z-index: 1;
+}
+
+.more-plant-card-img {
   opacity: 0;
   position: absolute;
   z-index: 2;
-  height: 400px;
-  overflow: hidden;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  font-size: 21px;
-  font-weight: 300;
-  cursor: pointer;
-  color: white;
 }
 
-.plant-card-see-more-2:hover {
+.more-plant-card:hover {
   opacity: .3;
 }
 
@@ -162,10 +74,7 @@
   font-size: 25px;
 }
 
-.fa-ex1 {
-  justify-items: flex-end;
-}
-
+/* check if this still needs to be h3 specific */
 h3.recent-plant-header {
   padding-top: 25px;
   padding-bottom: 25px;
